refactor(crd): update imports, labels, RBAC, and test references for kubernaut.ai group
Update all 18 Go import paths from api/apifrontend/apifrontend/v1alpha1 to
api/investigationsession/v1alpha1. Change label prefix from
apifrontend.kubernaut.ai/ to kubernaut.ai/ in session service constants and
test fixtures. Update RBAC apiGroup in deploy manifests, Helm templates, and
test infrastructure inline RBAC. Update CRD name assertions and apiVersion
strings in E2E and integration tests.

Closes #1223
