+++ deploy/apifrontend/base/02-rbac.yaml
@@ -8,10 +8,10 @@ metadata:
   name: apifrontend
 rules:
   # InvestigationSession controller (owned CRD)
-  - apiGroups: ["apifrontend.kubernaut.ai"]
+  - apiGroups: ["kubernaut.ai"]
     resources: ["investigationsessions"]
     verbs: ["get", "list", "watch", "create", "update", "patch", "delete"]
-  - apiGroups: ["apifrontend.kubernaut.ai"]
+  - apiGroups: ["kubernaut.ai"]
     resources: ["investigationsessions/status"]
     verbs: ["get", "update", "patch"]
   # RemediationRequest management (ADK A2A path — AF creates/lists on behalf of users)
